requirements.txt
-i https://pypi.org/simple
aiobotocore==0.12.0
aiohttp==3.6.2
aioitertools==0.5.1
amqp==2.5.2
appdirs==1.4.3
asgiref==3.2.3
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
billiard==3.6.3.0
black==19.10b0
boto3==1.12.15
botocore==1.15.15
celery==4.4.1
certifi==2019.11.28
chardet==3.0.4
click==7.0
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-cors-headers==3.2.1
django-filter==2.2.0
django-rest-framework==0.1.0
django-timezone-field==4.0
django==3.0.4
djangorestframework-simplejwt==4.4.0
djangorestframework==3.11.0
docutils==0.15.2
geojson==2.5.0
geomet==0.2.1.post1
gunicorn==20.0.4
html2text==2020.1.16
idna-ssl==1.1.0 ; python_version < '3.7'
idna==2.9
importlib-metadata==1.5.0
isort==4.3.21
jmespath==0.9.5
kombu==4.6.8
lazy-object-proxy==1.4.3
lxml==4.5.0
mccabe==0.6.1
multidict==4.7.5
pathspec==0.7.0
pillow==7.0.0
psycopg2==2.8.4
pyhumps==1.3.1
pyjwt==1.7.1
pylint==2.4.4
python-crontab==2.4.0
python-dateutil==2.8.1
pytz==2019.3
pyyaml==5.3
redis==3.4.1
regex==2020.2.20
requests==2.23.0
s3transfer==0.3.3
sentinelsat==0.13
shapely==1.7.0
six==1.14.0
sqlparse==0.3.1
tabulate==0.8.6
toml==0.10.0
tqdm==4.43.0
typed-ast==1.4.1
typing-extensions==3.7.4.2 ; python_version < '3.7'
urllib3==1.25.8
vine==1.3.0
wrapt==1.11.2
yarl==1.4.2
zipp==3.1.0
